cleanup: rename dissagg package to disagg
